Add null check on test cleanUp action

+++ src/test/java/redis/clients/jedis/commands/jedis/ClusterHotkeysCommandsTest.java
@@ -41,9 +41,11 @@ public class ClusterHotkeysCommandsTest extends ClusterJedisCommandsTestBase {
   }
 
   private void clearState() {
-    cluster.flushAll();
-    cluster.hotkeysStop();
-    cluster.hotkeysReset();
+    if (cluster != null) {
+      cluster.flushAll();
+      cluster.hotkeysStop();
+      cluster.hotkeysReset();
+    }
   }
 
   @Test
